% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{build_india_outline}
\alias{build_india_outline}
\title{Build an outline of India from state boundaries}
\usage{
build_india_outline(states)
}
\description{
Build an outline of India from state boundaries
}
\keyword{internal}
