% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{build_sidebar_html}
\alias{build_sidebar_html}
\title{Build the sidebar HTML/CSS/JS for the SpotMap}
\usage{
build_sidebar_html(
  n_cases,
  n_controls,
  mode,
  cluster_color,
  case_color,
  control_color
)
}
\description{
Build the sidebar HTML/CSS/JS for the SpotMap
}
\keyword{internal}
