% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{build_sidebar_js}
\alias{build_sidebar_js}
\title{Build the sidebar JavaScript for layer toggling}
\usage{
build_sidebar_js(cluster_color, case_color, control_color)
}
\description{
Build the sidebar JavaScript for layer toggling
}
\keyword{internal}
