% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{crop_geodataframe}
\alias{crop_geodataframe}
\title{Crop an sf object to a bounding box with margin}
\usage{
crop_geodataframe(sf_obj, bounds, margin = 1)
}
\description{
Crop an sf object to a bounding box with margin
}
\keyword{internal}
