% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{determine_mode}
\alias{determine_mode}
\title{Determine map mode: "districts", "states", or "india"}
\usage{
determine_mode(
  points_cases,
  district_name_col,
  state_name_col,
  count_cutoff = 2L
)
}
\description{
Determine map mode: "districts", "states", or "india"
}
\keyword{internal}
