% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loader.R
\name{load_data}
\alias{load_data}
\title{Read a data file (CSV, Excel, or TSV) into a data frame}
\usage{
load_data(path)
}
\arguments{
\item{path}{Path to data file.}
}
\value{
A data.frame.
}
\description{
Read a data file (CSV, Excel, or TSV) into a data frame
}
