% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{spatial_join}
\alias{spatial_join}
\title{Spatial join: attach state and district names to each point}
\usage{
spatial_join(
  df,
  lat_col,
  lon_col,
  states,
  districts,
  state_name_col,
  district_name_col
)
}
\description{
Spatial join: attach state and district names to each point
}
\keyword{internal}
